refactor: export router for polls and events
