+++ packages/portal/src/server-middleware/api/utils.js
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import defu  from 'defu';
 import createHttpError from 'http-errors';
 
@@ -35,14 +34,4 @@ export const forbiddenUnlessOriginAllowed = (origins) => (origin, callback) => {
   } else {
     callback(createHttpError(403, 'Origin not permitted'));
   }
-};
-
-export const keycloakUserinfo = async(req, config) => {
-  const keycloakUserinfoResponse = await axios({
-    baseURL: config.origin,
-    url: `/auth/realms/${config.realm}/protocol/openid-connect/userinfo`,
-    method: 'get',
-    headers: { authorization: req.headers.authorization }
-  });
-  return keycloakUserinfoResponse.data;
 };
